backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

psycopg[binary]==3.1.18

django-cors-headers==4.3.0
python-decouple==3.8

pyotp==2.9.0
qrcode[pil]==7.4.2

channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0

celery==5.3.4
redis==5.0.1

numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.2
transformers==4.35.2
sentence-transformers==2.2.2
# spacy==3.7.2  # Install separately with: python -m spacy download en_core_web_sm

# NLP Utilities
langdetect==1.0.9
textblob==0.17.1

# API Integrations
requests==2.31.0
geopy==2.4.1

# Document Generation
reportlab==4.0.9
icalendar==5.0.11
Pillow==10.1.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3

# Additional required packages
sqlparse==0.4.4